Separate Order admin in components
